C99 allows us to put %zu after all
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@9565 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ src/surf/surf_action.c
@@ -67,8 +67,8 @@ void *surf_action_new(size_t size, double cost, surf_model_t model,
                       int failed)
 {
   xbt_assert2(size <= action_mallocator_allocated_size,
-      "Cannot create a surf action of size %lu: the mallocator only provides actions of size %d",
-      (long unsigned) size, action_mallocator_allocated_size);
+      "Cannot create a surf action of size %zu: the mallocator only provides actions of size %d",
+      size, action_mallocator_allocated_size);
 
   surf_action_t action = xbt_mallocator_get(action_mallocator);
   action->refcount = 1;
